implement social cards
Social media cards are becoming increasingly more important for modern
websites and are displayed in multiple places e.g. on Twitter or in
Discord embeds.

The new cards module is capable of generating social cards with
specified title and description using a custom template. So far only
static cards for the most common pages are supported, although it should
be possible to render these live for dynamic content with minimal effort.

The social cards are saved in mikosite/media/cards directory.

To regenerate all static cards use the custom command:
python ./mikosite/manage.py updatecards

To add a new static social card for a new page:
1. Add the proper og and twitter properties pointing to the future
   location of the social card in the HTML template head:
   https://mikomath.org/media/cards/{PAGE_NAME}.png
2. Add a new entry to cards/static_cards.py:
   ("{PAGE_NAME}.png", "{PAGE_TITLE}", "{PAGE_DESCRIPTION}")
3. Run the command above to refresh all static social cards.