fix calendar popup
Commit 6fa2c3a6a2931f86c91d4e9b5d7489af68d953d1 modifed the calendar
popup to display the newly introduced information. It did not, however,
display correctly when some information was blank, causing 'None' to be
presented to the end user as if it was the missing value.

Modified the way data was passed to the JS script to correctly handle
blank values. Incidentally, fixed redundant spacing when no badges are
displayed.