enable template caching
The root and /kolo HTML templates now cache the dynamic content
(seminars and seminar-group) with a short expiry time to reduce the
number of calls to the database, which conserves the server's resources
and shortens the response time.

Expiry time of 5 or 15 minutes was used, which should not significantly
impact user experience.