fix bug with None in subtitle
Setting `null=True` for subtitle resulted in empty strings being replaced
with value of `None`. This caused the literal `None` to be displayed by
the template generator for all new posts which were missing a subtitle.

Modified the field to disallow null. Verified that empty subtitle is now
displayed properly.