dynamically load seminars on /kolo page
Previously, all seminars were inserted into the HTML template in the
form of a long JS script, which increased network load as well as
loading time and has caused subtle bugs in the past.
This will not be a sustainable model when we have hundreds or
thousands of seminars, most of them rarely viewed.

Removed the inefficient template generation and extended the JS script
to dynamically fetch relevant seminars when calendar page is changed
using the API introduced in 00c774d0cf7f055f9a80e4ecd31ba913f658f86b.