Add password validators
Enforced stronger constraints on passwords. We now
require 1 capital letter, 1 digit, and 1 special character.
Additionally, checks are run to make sure the password is different
from other user fields and is not on a blacklist (e.g. `password`).

Constraints are clearly communicated to the user when registering
or changing password in profile.